test(jsonld): adding behat test and move dummy to api resource
